Move Unfused and Fused attention to backends.py and some utils functions to pytorch/utils.py
