+++ transformer_engine/pytorch/distributed.py
@@ -565,7 +565,8 @@ def has_te_modules(network):
     """
     from .module import LayerNorm, RMSNorm
     from .module.base import TransformerEngineBaseModule
-    from .attention import UnfusedDotProductAttention, DotProductAttention
+    from .dot_product_attention.backends import UnfusedDotProductAttention
+    from .attention import DotProductAttention
     from .multi_head_attention import MultiheadAttention
     from .transformer import TransformerLayer
 
@@ -1479,7 +1480,8 @@ def _is_te_module(module):
     """
     from .module import LayerNorm, RMSNorm
     from .module.base import TransformerEngineBaseModule
-    from .attention import UnfusedDotProductAttention, DotProductAttention
+    from .dot_product_attention.backends import UnfusedDotProductAttention
+    from .attention import DotProductAttention
     from .multi_head_attention import MultiheadAttention
     from .transformer import TransformerLayer
